# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eudora Conseil & Relooking — API",
    "description": "API REST complète pour le site Eudora Conseil & Relooking.\n\n## Authentification\nL'API utilise **JWT (JSON Web Token)**. Pour accéder aux endpoints protégés :\n1. Appelez `POST /api/auth/login` avec vos identifiants\n2. Récupérez le `access_token` dans la réponse\n3. Ajoutez le header `Authorization: Bearer <access_token>` à chaque requête\n\n## Sections\n- **Auth** — Connexion, inscription, gestion du token\n- **CMS** — Gestion du contenu du site (Hero, About, Services, Formules, Témoignages, Galerie)\n- **Contact** — Messages du formulaire de contact\n- **Clients** — Gestion des clients\n- **Invoices** — Facturation\n- **Finance** — Dashboard financier et statistiques\n- **Settings** — Paramètres globaux du site",
    "version": "1.0.0",
    "contact": {
      "name": "Eudora Conseil & Relooking",
      "email": "info@ecrelooking.com",
      "url": "https://www.ecrelooking.com"
    }
  },
  "servers": [
    {
      "url": "https://www.ecrelooking.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:5000",
      "description": "Développement local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via POST /api/auth/login"
      }
    },
    "schemas": {
...
      "ContactMessage": {
        "type": "object",
        "required": ["name", "email", "message"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "example": "Marie Dupont" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string", "example": "+39 327 225 0364" },
          "service": { "type": "string", "example": "Colorimétrie" },
          "message": { "type": "string" },
          "is_read": { "type": "boolean", "default": false },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Client": {
        "type": "object",
        "required": ["first_name", "last_name", "email"],
        "properties": {
          "id": { "type": "integer" },
          "uuid": { "type": "string", "format": "uuid" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "full_name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "country": { "type": "string", "default": "France" },
          "notes": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive", "prospect"] },
          "preferred_language": { "type": "string", "enum": ["fr", "en", "it"] },
          "total_spent": { "type": "number" },
          "invoices_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
...
      "pageParam": {
        "name": "page",
        "in": "query",
        "description": "Numéro de page",
        "schema": { "type": "integer", "default": 1 }
      },
      "perPageParam": {
        "name": "per_page",
        "in": "query",
        "description": "Éléments par page",
        "schema": { "type": "integer", "default": 20 }
      }
    }
  },
  "paths": {
...
    "/api/contact/": {
      "post": {
        "tags": ["Contact"],
        "summary": "Envoyer un message (public)",
        "description": "Endpoint utilisé par le formulaire de contact du site. Sauvegarde en base + notification email à l'admin.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "email", "message"], "properties": { "name": { "type": "string" }, "email": { "type": "string", "format": "email" }, "phone": { "type": "string" }, "service": { "type": "string" }, "message": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Message envoyé" }, "400": { "description": "Champs manquants ou email invalide" } }
      },
      "get": {
        "tags": ["Contact"],
        "summary": "Lister les messages (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/pageParam" }, { "$ref": "#/components/parameters/perPageParam" }, { "name": "unread", "in": "query", "schema": { "type": "boolean" } }],
        "responses": { "200": { "description": "Liste paginée des messages", "content": { "application/json": { "schema": { "type": "object", "properties": { "messages": { "type": "array", "items": { "$ref": "#/components/schemas/ContactMessage" } }, "total": { "type": "integer" }, "page": { "type": "integer" }, "pages": { "type": "integer" }, "unread_count": { "type": "integer" } } } } } } }
      }
    },
    "/api/contact/{id}": {
      "get": {
        "tags": ["Contact"],
        "summary": "Voir un message (le marque comme lu)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Détail du message" }, "404": { "description": "Message introuvable" } }
      },
      "delete": {
        "tags": ["Contact"],
        "summary": "Supprimer un message",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Message supprimé" } }
      }
    },
    "/api/clients/": {
      "get": {
        "tags": ["Clients"],
        "summary": "Lister les clients",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/pageParam" }, { "$ref": "#/components/parameters/perPageParam" }, { "name": "search", "in": "query", "schema": { "type": "string" } }, { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "inactive", "prospect"] } }],
        "responses": { "200": { "description": "Liste paginée des clients" } }
      },
      "post": {
        "tags": ["Clients"],
        "summary": "Créer un client",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } } },
        "responses": { "201": { "description": "Client créé" }, "409": { "description": "Email déjà existant" } }
      }
    },
    "/api/clients/{id}": {
      "get": {
        "tags": ["Clients"],
        "summary": "Voir un client",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Détail du client" } }
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Modifier un client",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } } },
        "responses": { "200": { "description": "Client modifié" } }
      },
      "delete": {
        "tags": ["Clients"],
        "summary": "Supprimer un client",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Client supprimé" } }
      }
    },
    "/api/invoices/": {
      "get": {
        "tags": ["Invoices"],
        "summary": "Lister les factures",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/pageParam" }, { "$ref": "#/components/parameters/perPageParam" }, { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "sent", "paid", "overdue", "cancelled"] } }, { "name": "client_id", "in": "query", "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Liste paginée des factures" } }
      },
      "post": {
        "tags": ["Invoices"],
        "summary": "Créer une facture",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["client_id"], "properties": { "client_id": { "type": "integer" }, "notes": { "type": "string" }, "tva_rate": { "type": "number", "default": 0.2 }, "items": { "type": "array", "items": { "type": "object", "required": ["description", "unit_price"], "properties": { "description": { "type": "string" }, "quantity": { "type": "number", "default": 1 }, "unit_price": { "type": "number" } } } } } } } } },
        "responses": { "201": { "description": "Facture créée avec numéro auto-généré (EUD-YYYYMM-XXXX)" } }
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "tags": ["Invoices"],
        "summary": "Voir une facture",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Détail de la facture avec lignes" } }
      },
      "put": {
        "tags": ["Invoices"],
        "summary": "Modifier une facture (statut, paiement)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string", "enum": ["draft", "sent", "paid", "overdue", "cancelled"] }, "payment_method": { "type": "string" }, "notes": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Facture modifiée" } }
      },
      "delete": {
        "tags": ["Invoices"],
        "summary": "Supprimer une facture (brouillon uniquement)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Facture supprimée" }, "400": { "description": "Seuls les brouillons peuvent être supprimés" } }
      }
    },
...
  }
}